% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validity.R
\name{reg}
\alias{reg}
\title{The default OLS regression}
\usage{
reg(y, X)
}
\arguments{
\item{y}{Dependent variable, vector of (nx1)}

\item{X}{Regressors, can be either a vector (nx1) or a matrix (nxm) if there is more than one regressor}
}
\value{
The function returns the fitted values based on the input and the use of standard OLS (including a constant)
}
\description{
This function generates fitted values based on a set of regressors X and the independent variable y.
}
\examples{
x <- rnorm(100, mean=0, sd=1)
y <- 10*x
reg(y,x)
}
\author{
Florian Schuetze 26.08.2024
}
