% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validity.R
\name{Example1}
\alias{Example1}
\title{The first example of Frahm (2023)}
\usage{
Example1(c, n, tau, N)
}
\arguments{
\item{c}{A constant}

\item{n}{The number of observations}

\item{tau}{(optional, default is tau = 1) tau is the standard deviation of the error e in the regression Y = g(X) + e}

\item{N}{(optional, default is N = 1000) N is the number of bootstrap replications in the validity function}
}
\value{
The function returns the p-value of the validity test.
}
\description{
This function replicates the first example of Frahm (2023)
}
\examples{
Example1(0.25,100)
}
\references{
Frahm, G., 2023, A Test for the Validity of Regression Models. Available on SSRN: https://ssrn.com/abstract=4610329
}
\author{
Florian Schuetze 26.08.2024
}
