% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validity.R
\name{Example2}
\alias{Example2}
\title{The second example of Frahm (2023)}
\usage{
Example2(c, n)
}
\arguments{
\item{c}{A constant}

\item{n}{The number of observations}
}
\value{
The function returns the p-value of the validity test.
}
\description{
This function replicates the second example of Frahm (2023)
}
\examples{
Example2(0.25,100)
}
\references{
Frahm, G., 2023, A Test for the Validity of Regression Models. Available on SSRN: https://ssrn.com/abstract=4610329
}
\author{
Florian Schuetze 26.08.2024
}
