% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validity.R
\name{Example3}
\alias{Example3}
\title{The third example of Frahm (2023)}
\usage{
Example3(c, n)
}
\arguments{
\item{c}{A constant}

\item{n}{The number of observations}
}
\value{
The function returns the p-value of the validity test.
}
\description{
This function replicates the third example of Frahm (2023)
}
\examples{
Example3(0.25,100)
}
\references{

}
\author{
Florian Schuetze 26.08.2024
}
